+++ udata/core/topic/apiv2.py
@@ -67,6 +67,8 @@ topic_fields = apiv2.model('Topic', {
     'spatial': fields.Nested(
         spatial_coverage_fields, allow_null=True,
         description='The spatial coverage'),
+    'last_modified': fields.ISODateTime(
+        description='The topic last modification date', readonly=True),
     'organization': fields.Nested(
         org_ref_fields, allow_null=True,
         description='The publishing organization', readonly=True),